feat: IDX-25 IVF-SQ8 + BIN-02 IVF-FLAT
IDX-25: IVF-SQ8 Complete Integration
- Add IndexType::IvfSq8 with from_str() support
- Add IndexParams::ivf_sq8() constructor
- Codec serialization (type ID: 6)
- Full FFI C API support (CIndexType, IndexWrapper)
- 8 tests pass, 349 total tests pass

BIN-02: BIN_IVF_FLAT Initial Implementation
- Add Binary IVF Flat index (src/faiss/bin_ivf_flat.rs)
- Hamming distance for binary vectors
- IVF clustering for binary data

cargo check: OK (179 warnings, 0 errors)

+++ src/faiss/sparse_inverted_cc.rs
@@ -551,8 +551,8 @@ mod tests {
             let index_clone = Arc::clone(&index);
             let handle = thread::spawn(move || {
                 for j in 0..10 {
-                    let vec = SparseVector::from_pairs(&[(i, 1.0), (j, 2.0)]);
-                    index_clone.add(&vec, i * 10 + j).unwrap();
+                    let vec = SparseVector::from_pairs(&[(i as u32, 1.0), (j as u32, 2.0)]);
+                    index_clone.add(&vec, (i * 10 + j) as i64).unwrap();
                 }
             });
             handles.push(handle);
@@ -574,7 +574,7 @@ mod tests {
         // 先添加一些数据
         for i in 0..100 {
             let vec = SparseVector::from_pairs(&[(i % 10, 1.0), (((i + 5) % 10), 2.0)]);
-            index.add(&vec, i).unwrap();
+            index.add(&vec, i as i64).unwrap();
         }
         
         // 并发搜索
@@ -602,7 +602,7 @@ mod tests {
         let mut vectors = Vec::new();
         for i in 0..100 {
             let vec = SparseVector::from_pairs(&[(i % 10, 1.0)]);
-            vectors.push((i, vec));
+            vectors.push((i as i64, vec));
         }
         
         index.add_batch(&vectors).unwrap();
